Database fixes and improvements
* changed the id type of the room db entities from Int to Long;
* fixed some issues related to the incorrect outfit editing and adding.

+++ app/src/main/java/com/olehmaliuta/clothesadvisor/api/http/requests/UploadOutfitRequest.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class UploadOutfitRequest (
     @SerializedName("name") var name: String,
-    @SerializedName("item_ids") var itemIds: List<Int>
+    @SerializedName("item_ids") var itemIds: List<Long>
 )
