+++ app/src/main/java/com/olehmaliuta/clothesadvisor/api/http/responses/CombinationResponse.kt
@@ -3,7 +3,7 @@ package com.olehmaliuta.clothesadvisor.api.http.responses
 import com.google.gson.annotations.SerializedName
 
 data class CombinationResponse (
-    @SerializedName("id") var id: Int? = null,
+    @SerializedName("id") var id: Long? = null,
     @SerializedName("name") var name: String? = null,
-    @SerializedName("items") var items: List<Int>? = null,
+    @SerializedName("items") var items: List<Long>? = null,
 )
