+++ app/src/main/java/com/olehmaliuta/clothesadvisor/api/http/responses/NewCombinationResponse.kt
@@ -3,5 +3,5 @@ package com.olehmaliuta.clothesadvisor.api.http.responses
 import com.google.gson.annotations.SerializedName
 
 data class NewCombinationResponse (
-    @SerializedName("combination_id") var combinationId: Int? = null,
+    @SerializedName("combination_id") var combinationId: Long? = null,
 )
