+++ app/src/main/java/com/olehmaliuta/clothesadvisor/api/http/responses/UserProfileResponse.kt
@@ -3,6 +3,6 @@ package com.olehmaliuta.clothesadvisor.api.http.responses
 import com.google.gson.annotations.SerializedName
 
 data class UserProfileResponse (
-    @SerializedName("id") var id: Int? = null,
+    @SerializedName("id") var id: Long? = null,
     @SerializedName("email") var email: String? = null,
 )
