+++ app/src/main/java/com/olehmaliuta/clothesadvisor/database/entities/query/OutfitWithClothingItemCount.kt
@@ -3,7 +3,7 @@ package com.olehmaliuta.clothesadvisor.database.entities.query
 import androidx.room.ColumnInfo
 
 data class OutfitWithClothingItemCount (
-    @ColumnInfo(name = "id") val id: Int,
+    @ColumnInfo(name = "id") val id: Long,
     @ColumnInfo(name = "name") val name: String,
-    @ColumnInfo(name = "item_count") val itemCount: Int
+    @ColumnInfo(name = "item_count") val itemCount: Long
 )
