+++ app/src/main/java/com/olehmaliuta/clothesadvisor/database/entities/query/OutfitWithClothingItemIds.kt
@@ -3,7 +3,7 @@ package com.olehmaliuta.clothesadvisor.database.entities.query
 import com.google.gson.annotations.SerializedName
 
 data class OutfitWithClothingItemIds(
-    @SerializedName("id") val id: Int,
+    @SerializedName("id") val id: Long,
     @SerializedName("name") val name: String,
-    @SerializedName("item_ids") val itemIds: List<Int>
+    @SerializedName("item_ids") val itemIds: List<Long>
 )
